Mostly fixes and optimizations.

+++ src/main/java/me/TechsCode/TechDiscordBot/modules/VerificationChannel.java
@@ -58,15 +58,15 @@ public class VerificationChannel extends Module {
                 if(bot.getSpigotAPI().isAvailable()) {
                     if(apiNotAvailable != null) {
                         apiNotAvailable = null;
-                        //sendInstructions();
+                        sendInstructions();
                     }
                 } else {
                     if(apiNotAvailable == null) {
-                        //lastInstructions.delete().complete();
-                        //CustomEmbedBuilder message = new CustomEmbedBuilder("Verification Disabled")
-                        //        .setText("The Web API is currently unavailable. Please contact staff for more info!")
-                        //        .error();
-                        //apiNotAvailable = message.send(channel);
+                        //(lastInstructions != null) lastInstructions.delete().complete();
+                        CustomEmbedBuilder message = new CustomEmbedBuilder()
+                                .setText("The Web API is currently unavailable. You cannot verify until it's online again!\n**Sorry for another inconvenience!**")
+                                .error();
+                        apiNotAvailable = message.send(channel);
                     }
                 }
                 try {
@@ -80,9 +80,8 @@ public class VerificationChannel extends Module {
 
     @Override
     public void onDisable() {
-        if(lastInstructions != null) {
-            lastInstructions.delete().submit();
-        }
+        if(lastInstructions != null) lastInstructions.delete().submit();
+        if(apiNotAvailable != null) apiNotAvailable.delete().complete();
     }
 
     @Override
@@ -100,10 +99,7 @@ public class VerificationChannel extends Module {
     @SubscribeEvent
     public void onMessage(GuildMessageReceivedEvent e) {
         if(e.getAuthor().isBot()) return;
-
-        if(!e.getChannel().equals(VERIFICATION_CHANNEL.query().first())) {
-            return;
-        }
+        if(!e.getChannel().equals(VERIFICATION_CHANNEL.query().first())) return;
 
         if(true) {// Return it for now ;p
             e.getMessage().delete().queue();
@@ -183,11 +179,13 @@ public class VerificationChannel extends Module {
     public void sendInstructions() {
         CustomEmbedBuilder howItWorksMessage = new CustomEmbedBuilder("Verification is Broken")
                 .setText("**How do I do it now?**\nPlease let a Staff Member know that you would like to verify. You can either DM them or mention one of them in #general. When you do so, make sure to give them a link to your spigot profile. Please also post `TechVerification` on it so we can verify you own the account. Hopefully soon, a staff member will get back to you telling you that your account is linked. Shortly after, you should get your roles!\n\n**Sorry for the inconvenience!**");
-        lastInstructions = howItWorksMessage.send(channel);
-        //if(apiNotAvailable != null) apiNotAvailable.delete().complete();
-        //if(lastInstructions != null) lastInstructions.delete().complete();
+        if(apiNotAvailable != null && bot.getSpigotAPI().isAvailable()) {
+            apiNotAvailable.delete().complete();
+            apiNotAvailable = null;
+        }
+        if(lastInstructions != null) lastInstructions.delete().complete();
         //CustomEmbedBuilder howItWorksMessage = new CustomEmbedBuilder("How It Works")
         //        .setText("Type your SpigotMC name in this Chat to verify.");
-        //lastInstructions = howItWorksMessage.send(channel);
+        lastInstructions = howItWorksMessage.send(channel);
     }
 }
